remove redux devtools from router reducer
