+++ packages/next/src/shared/lib/router/action-queue.ts
@@ -7,7 +7,6 @@ import {
   ACTION_NAVIGATE,
   ACTION_RESTORE,
 } from '../../../client/components/router-reducer/router-reducer-types'
-import type { ReduxDevToolsInstance } from '../../../client/components/use-reducer-with-devtools'
 import { reducer } from '../../../client/components/router-reducer/router-reducer'
 import { startTransition } from 'react'
 import { isThenable } from '../is-thenable'
@@ -16,7 +15,6 @@ export type DispatchStatePromise = React.Dispatch<ReducerState>
 
 export type AppRouterActionQueue = {
   state: AppRouterState
-  devToolsInstance?: ReduxDevToolsInstance
   dispatch: (payload: ReducerActions, setState: DispatchStatePromise) => void
   action: (state: AppRouterState, action: ReducerActions) => ReducerState
   pending: ActionQueueNode | null
@@ -84,10 +82,6 @@ async function runAction({
     }
 
     actionQueue.state = nextState
-
-    if (actionQueue.devToolsInstance) {
-      actionQueue.devToolsInstance.send(payload, nextState)
-    }
 
     runRemainingActions(actionQueue, setState)
     action.resolve(nextState)
